requirements.txt
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure-pytest
requests~=2.28.1
PyYAML~=6.0
pytest~=7.1.2
selenium~=4.3.0
Appium-Python-Client~=2.6.0